More pace fixes and optimisations

+++ RunAge/app/src/main/java/xevenition/com/runage/util/CalorieCalculator.kt
@@ -8,7 +8,7 @@ object CalorieCalculator {
      * @param weight in kilograms
      * calculation based on https://fitness.stackexchange.com/a/36045
      */
-    fun getCaloriesBurned(distance: Double, weight: Float): Double {
-        return distance * weight *  1.036 / 1000
+    fun getCaloriesBurned(distance: Double, weight: Float): Int {
+        return (distance * weight *  1.036 / 1000).toInt()
     }
 }
